Extract `TensorHook` from `FrameworkHook`
